feat: Add console.log statement to display response message in submitUserMessage
The code changes in `actions.tsx` add a console.log statement to display the response message received after submitting the user's message. This enhances the functionality of the `submitUserMessage` function by providing immediate feedback to the user.

I forgot to promote the current deployment, it's already fixed a long time ago.